download file support offset and download bytes
git-svn-id: http://fastdfs.googlecode.com/svn@276 0d4c4929-bc51-0410-a161-b7c81ae5c018

+++ trunk/common/logger.c
@@ -178,7 +178,7 @@ static void doLog(const char *caption, const char* text, const int text_len, \
 	{
 		fprintf(stderr, "file: "__FILE__", line: %d, " \
 			"log buff size: %d < log text length: %d ", \
-			__LINE__, sizeof(log_buff), text_len + 64);
+			__LINE__, (int)sizeof(log_buff), text_len + 64);
 		pthread_mutex_unlock(&log_thread_lock);
 		return;
 	}
